@@ -14,26 +14,32 @@
   "scripts": {
     "build": "vite build && npm run copy",
     "watch": "npm-watch",
-    "copy": "node ./scripts/copyAssets.ts",
-    "prettier": "prettier --write ."
+    "copy": "esno ./scripts/copyAssets.ts",
+    "lint": "eslint --cache ."
   },
   "devDependencies": {
-    "@iconify-json/carbon": "^1.1.16",
-    "@types/chrome": "^0.0.231",
-    "@types/node": "^18.15.11",
-    "@vitejs/plugin-vue": "^4.1.0",
-    "@vueuse/core": "^9.13.0",
+    "@antfu/eslint-config": "^0.39.8",
+    "@iconify-json/carbon": "^1.1.18",
+    "@types/chrome": "^0.0.242",
+    "@types/node": "^20.4.5",
+    "@typescript-eslint/eslint-plugin": "^6.2.0",
+    "@typescript-eslint/parser": "^6.2.0",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vueuse/core": "^10.3.0",
     "cheerio": "^1.0.0-rc.12",
-    "element-plus": "^2.3.3",
-    "fast-glob": "^3.2.12",
+    "dayjs": "^1.11.9",
+    "element-plus": "^2.3.8",
+    "eslint": "^8.46.0",
+    "esno": "^0.17.0",
+    "fast-glob": "^3.3.1",
     "npm-watch": "^0.11.0",
-    "pinia": "^2.0.34",
-    "prettier": "^2.8.7",
-    "unocss": "^0.51.4",
-    "unplugin-auto-import": "^0.15.3",
-    "unplugin-icons": "^0.16.1",
-    "unplugin-vue-components": "^0.24.1",
-    "vite": "^4.2.1",
-    "vue": "^3.2.47"
+    "sass": "^1.64.1",
+    "typescript": "^5.1.6",
+    "unocss": "^0.54.0",
+    "unplugin-auto-import": "^0.16.6",
+    "unplugin-icons": "^0.16.5",
+    "unplugin-vue-components": "^0.25.1",
+    "vite": "^4.4.7",
+    "vue": "^3.3.4"
   }
 }
