@@ -1,2 +1,3 @@
-export const getImgSrc = (prop: string) =>
-  new URL(`/src/assets/image/${prop}`, import.meta.url).href
+export function getImgSrc(prop: string, type: string = 'common') {
+  return new URL(`/src/assets/image/${type}/${prop}.png`, import.meta.url).href
+}
