@@ -1,6 +1,5 @@
 import { createApp } from 'vue'
 import Main from './main.vue'
-import { createPinia } from 'pinia'
-import '@/styles'
+import '~/styles'
 
-createApp(Main).use(createPinia()).mount('#app')
+createApp(Main).mount('#app')
