@@ -1,3 +1,7 @@
+<script setup lang="ts">
+defineProps(['tableData'])
+</script>
+
 <template>
   <el-table :data="tableData">
     <el-table-column type="expand">
@@ -6,16 +10,17 @@
           <div class="prt-reward-item">
             <div class="prt-item-list">
               <div
+                v-for="treasure, idx in row.treasureList"
+                :key="idx"
                 class="lis-treasure btn-treasure-item"
-                v-for="treasure in row.treasureList"
               >
                 <div class="prt-treasure-image">
-                  <img class="img-treasure-item" :src="treasure.src" />
-                  <div class="prt-article-count" v-if="treasure.number">
+                  <img class="img-treasure-item" :src="treasure.src">
+                  <div v-if="treasure.number" class="prt-article-count">
                     <span class="txt-small">x</span>{{ treasure.number }}
                   </div>
                 </div>
-                <div :class="treasure.boxClass"></div>
+                <div :class="treasure.boxClass" />
               </div>
             </div>
           </div>
@@ -30,7 +35,3 @@
     <el-table-column prop="speed" label="跑速" align="center" width="100" />
   </el-table>
 </template>
-
-<script setup lang="ts">
-defineProps(['tableData'])
-</script>
