@@ -1,31 +1,17 @@
 <template>
-  <div class="fc flex-col">
-    <div
-      text-10
-      class="icon-btn"
-      i-carbon-sun
-      dark:i-carbon-moon
-      @click="toggleDark()"
-    ></div>
+  <div fc flex-col>
+    <div text-10 icon-btn i-carbon-sun dark:i-carbon-moon @click="toggleDark()" />
     <a
-      text-5
-      m-2
-      i-carbon:debug
-      hover:text-amber
+      text-5 m-2 i-carbon:debug hover:text-amber
       rel="noreferrer"
       href="https://bbs.nga.cn/read.php?tid=30999435"
-      target="_blank"
-      title="报Bug"
+      target="_blank" title="报Bug"
     />
     <a
-      text-5
-      m-2
-      i-carbon:logo-github
-      hover:text-amber
+      text-5 m-2 i-carbon:logo-github hover:text-amber
       rel="noreferrer"
       href="https://github.com/Waaatanuki/Chrome-Extension-Tarou"
-      target="_blank"
-      title="GitHub"
+      target="_blank" title="GitHub"
     />
   </div>
 </template>
