@@ -1,16 +1,3 @@
-<template>
-  <div class="flex flex-wrap">
-    <ToDoList />
-    <StoneCount />
-    <RecoveryItemTable />
-    <Contact class="absolute right-15px top-15px" />
-  </div>
-</template>
-
-<script lang="ts">
-export default { name: 'Dashboard' }
-</script>
-
 <script setup lang="ts">
 import ToDoList from './components/ToDoList.vue'
 import StoneCount from './components/StoneCount.vue'
@@ -18,4 +5,11 @@ import RecoveryItemTable from './components/RecoveryItemTable.vue'
 import Contact from './components/Contact.vue'
 </script>
 
-<style lang="scss" scoped></style>
+<template>
+  <div flex flex-wrap>
+    <ToDoList />
+    <StoneCount />
+    <RecoveryItemTable />
+    <Contact absolute right-15px top-15px />
+  </div>
+</template>
