@@ -1,75 +1,77 @@
+<script setup lang="ts">
+import type { EvokerInfo } from '~/logic/types'
+import { domainList, evokerList, tarotList, weaponList } from '~/constants/evoker'
+
+const evoker = defineModel<EvokerInfo>({ required: true })
+</script>
+
 <template>
-  <el-card
-    class="evoker-card w-130px"
-    :body-style="{ padding: '0px' }"
-    shadow="hover"
-  >
-    <img
-      w-130px
-      :src="`https://prd-game-a1-granbluefantasy.akamaized.net/assets/img/sp/assets/npc/m/${evokerInfo.npcId}_01.jpg`"
-    />
+  <el-card class="evoker-card w-130px" :body-style="{ padding: '0px' }" shadow="hover">
+    <img w-130px :src="`https://prd-game-a1-granbluefantasy.akamaized.net/assets/img/sp/assets/npc/m/${evoker.npcId}_01.jpg`">
     <div class="h-170px flex flex-col items-center justify-evenly">
       <el-select
-        v-model="evokerInfo.tarotLevel"
+        v-model="evoker.tarotLevel"
         class="w-100px"
-        :class="{ completed: evokerInfo.tarotLevel == tarotList.at(-1)?.value }"
+        :class="{ completed: evoker.tarotLevel === tarotList.at(-1)?.value }"
         size="small"
         placeholder="塔罗牌进度"
         placement="bottom"
       >
         <el-option
-          v-for="item in tarotList"
+          v-for="item, idx in tarotList"
+          :key="idx"
           :label="item.label"
           :value="item.value"
         />
       </el-select>
 
       <el-select
-        v-model="evokerInfo.evokerLevel"
+        v-model="evoker.evokerLevel"
         class="w-100px"
-        :class="{
-          completed: evokerInfo.evokerLevel == evokerList.at(-1)?.value,
-        }"
+        :class="{ completed: evoker.evokerLevel === evokerList.at(-1)?.value }"
         size="small"
         placeholder="贤者进度"
         placement="bottom"
       >
         <el-option
-          v-for="item in evokerList"
+          v-for="item, idx in evokerList"
+          :key="idx"
           :label="item.label"
           :value="item.value"
         />
       </el-select>
 
       <el-select
-        v-model="evokerInfo.weaponLevel"
+        v-model="evoker.weaponLevel"
         class="w-100px"
         :class="{
-          completed: evokerInfo.weaponLevel == weaponList.at(-1)?.value,
+          completed: evoker.weaponLevel === weaponList.at(-1)?.value,
         }"
         size="small"
         placeholder="贤武进度"
         placement="bottom"
       >
         <el-option
-          v-for="item in weaponList"
+          v-for="item, idx in weaponList"
+          :key="idx"
           :label="item.label"
           :value="item.value"
         />
       </el-select>
 
       <el-select
-        v-model="evokerInfo.domainLevel"
+        v-model="evoker.domainLevel"
         class="w-100px"
         :class="{
-          completed: evokerInfo.domainLevel == domainList.at(-1)?.value,
+          completed: evoker.domainLevel === domainList.at(-1)?.value,
         }"
         size="small"
         placeholder="领域进度"
         placement="bottom"
       >
         <el-option
-          v-for="item in domainList"
+          v-for="item, idx in domainList"
+          :key="idx"
           :label="item.label"
           :value="item.value"
         />
@@ -77,28 +79,18 @@
 
       <div>
         <el-checkbox
-          v-model="evokerInfo.isAbility4Release"
+          v-model="evoker.isAbility4Release"
           label="四技能解锁"
           size="small"
         />
       </div>
 
       <div>
-        <el-checkbox v-model="evokerInfo.target" label="目标" size="small" />
+        <el-checkbox v-model="evoker.target" label="目标" size="small" />
       </div>
     </div>
   </el-card>
 </template>
-
-<script setup lang="ts">
-import {
-  tarotList,
-  evokerList,
-  weaponList,
-  domainList,
-} from '@/settings/evoker.setting'
-defineProps(['evokerInfo'])
-</script>
 
 <style lang="scss">
 .evoker-card {
