@@ -1,12 +1,5 @@
-<template>
-  <div
-    class="icon-btn"
-    i-carbon-sun
-    dark:i-carbon-moon
-    @click="toggleDark()"
-  ></div>
-</template>
-
 <script setup lang="ts"></script>
 
-<style scoped></style>
+<template>
+  <div class="icon-btn" i-carbon-sun dark:i-carbon-moon @click="toggleDark()" />
+</template>
